politica de producto y pago de producto

+++ app/Http/Controllers/ProductoController.php
@@ -73,7 +73,6 @@ class ProductoController extends Controller
      */
     public function create()
     {
-        $this->authorize('create');
         $categorias = Categoria::all();
         return view("supervisor.producto.create",compact('categorias'));
     }
@@ -105,6 +104,7 @@ class ProductoController extends Controller
      */
     public function show($id)
     {
+
         $producto = Producto::find($id);
         return view ("supervisor.producto.show",compact('producto'));
 
@@ -120,6 +120,7 @@ class ProductoController extends Controller
     {
         $categorias = Categoria::all();
         $producto = Producto::find($id);
+        $this->authorize('edit',$producto);
         return view ("supervisor.producto.edit",compact('producto','categorias'));
     }
 
@@ -153,6 +154,7 @@ class ProductoController extends Controller
     public function destroy($id)
     {
         $producto = producto::findOrFail($id);
+         $this->authorize('destroy',$producto);
         Storage::delete('public/'.$producto->imagen);
         producto::destroy($id);
         return redirect('/producto');
